+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Migrate to latest v2 mongo driver [#127](https://github.com/rokwire/lms-building-block/issues/127)
 
 ## [1.15.2] - 2026-07-07
 ### Added